# internal/routes/openapi.yaml
openapi: 3.0.3
info:
  title: Go Gin Boilerplate API
  version: 1.0.0
  description: REST endpoints for authentication and RBAC. GraphQL has its own schema.
servers:
  - url: /
paths:
...
  /api/v1/rbac/users:
    get:
      tags: [Users]
      summary: List users
      security: [{ bearerAuth: [] }]
      x-required-permission: user:index
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PerPage' }
        - { $ref: '#/components/parameters/Search' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/v1/rbac/users/{id}/roles:
    parameters:
      - { $ref: '#/components/parameters/ID' }
    get:
      tags: [Users]
      summary: Get roles assigned to a user
      security: [{ bearerAuth: [] }]
      x-required-permission: user:assign_role
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Users]
      summary: Assign roles to a user
      security: [{ bearerAuth: [] }]
      x-required-permission: user:assign_role
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AssignRolesInput' }
      responses:
        '200': { $ref: '#/components/responses/Success' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ID:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Search:
      name: search
      in: query
      schema: { type: string, maxLength: 100 }
  responses:
    Success:
      description: Successful response
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Envelope' }
...
    Forbidden:
      description: Missing permission
...
  schemas:
    Envelope:
      type: object
      required: [status, status_code, request_id, message, data, extra_data, errors]
      properties:
        status: { type: boolean, description: True for a successful response }
        status_code: { type: integer, description: Application status or error code }
        request_id: { type: string, format: uuid }
        message: { type: string, description: Localized human-readable message }
        data: {}
        extra_data: {}
        errors: {}
...
    AssignRolesInput:
      type: object
      required: [role_ids]
      properties:
        role_ids:
          type: array
          items: { type: integer, minimum: 1 }
...